Split entry model field refs, OTP auto-type, and trash history into mixins.
